Move Serializer into its own subdirectory
